requirements.txt
black==24.4.2
click==8.1.7
fonttools==4.53.0
mypy-extensions==1.0.0
numpy==1.26.4
packaging==24.1
pathspec==0.12.1
pillow==10.3.0
platformdirs==4.2.2
pypinyin==0.51.0
termcolor==2.4.0
